feat: split mobile and desktop settings
